Cache control completeness
