+++ okhttp/src/nonJvmMain/kotlin/okhttp3/CacheControl.kt
@@ -17,6 +17,7 @@ package okhttp3
 
 import kotlin.time.DurationUnit
 import okhttp3.internal.commonBuild
+import okhttp3.internal.commonForceCache
 import okhttp3.internal.commonForceNetwork
 import okhttp3.internal.commonImmutable
 import okhttp3.internal.commonMaxAge
@@ -26,6 +27,7 @@ import okhttp3.internal.commonNoCache
 import okhttp3.internal.commonNoStore
 import okhttp3.internal.commonNoTransform
 import okhttp3.internal.commonOnlyIfCached
+import okhttp3.internal.commonParse
 import okhttp3.internal.commonToString
 
 actual class CacheControl internal actual constructor(
@@ -76,5 +78,9 @@ actual class CacheControl internal actual constructor(
 
   actual companion object {
     actual val FORCE_NETWORK = commonForceNetwork()
+
+    actual val FORCE_CACHE = commonForceCache()
+
+    actual fun parse(headers: Headers): CacheControl = commonParse(headers)
   }
 }
